requirements.txt
gmqtt==0.6.11
PyYAML==5.3
voluptuous==0.11.7
pytest==8.3.5